fixes for poné sound + cleaning

+++ d_bankp.c
@@ -12,8 +12,8 @@
 
 int pcm1=-1;
 Sint16 *nSoundBuffer=NULL;
-extern unsigned int frame_x;
-extern unsigned int frame_y;
+extern unsigned short frame_x;
+extern unsigned short frame_y;
 #endif
  
 int ovlInit(char *szShortName)
