+++ d_freekick.c
@@ -10,8 +10,8 @@
 
 int pcm1=-1;
 Sint16 *nSoundBuffer=NULL;
-extern unsigned int frame_x;
-extern unsigned int frame_y;
+extern unsigned short frame_x;
+extern unsigned short frame_y;
 #else
 static void Set6PCM();
 void PCM_MeStop(PcmHn hn);
