+++ d_segae.c
@@ -14,8 +14,8 @@
 
 int pcm1=-1;
 Sint16 *nSoundBuffer=NULL;
-extern unsigned int frame_x;
-extern unsigned int frame_y;
+extern unsigned short frame_x;
+extern unsigned short frame_y;
 //UINT16 map[0x1000];
 #endif
